Restored parallel SILO
