+++ build_octotiger.sh
@@ -16,7 +16,7 @@ rm -r CMakeFiles
 
 
 cmake -DCMAKE_PREFIX_PATH="$HOME/local/$1/hpx" -DCMAKE_CXX_FLAGS="-DBOOST_USE_VALGRIND" \
-      -DCMAKE_CXX_COMPILER=g++ -DOCTOTIGER_WITH_TESTS=OFF \
+      -DCMAKE_CXX_COMPILER=g++ -DOCTOTIGER_WITH_TESTS=ON \
       -DCMAKE_C_COMPILER=gcc \
       -DCMAKE_CXX_FLAGS="-pg -DBOOST_USE_VALGRIND -L$HOME/local/boost/lib -march=native" \
       -DCMAKE_C_FLAGS="-L$HOME/local/boost/lib" \
