+++ octotiger/silo.hpp
@@ -9,6 +9,7 @@
 #include "octotiger/config/export_definitions.hpp"
 #include "octotiger/defs.hpp"
 #include "octotiger/real.hpp"
+#include "octotiger/node_location.hpp"
 
 #include <hpx/include/naming.hpp>
 
@@ -17,7 +18,14 @@
 #include <utility>
 #include <vector>
 
+
+
 #include <silo.h>
+
+
+#define SILO_DRIVER DB_HDF5
+#define SILO_VERSION 110
+
 
 class node_server;
 
@@ -80,5 +88,53 @@ void output_all(std::string fname, int cycle, bool);
 void load_options_from_silo(std::string fname, DBfile* = nullptr);
 
 OCTOTIGER_EXPORT void load_data_from_silo(std::string fname, node_server*, hpx::id_type);
+
+
+template<class T>
+struct db_type {
+	static constexpr int d = DB_UNKNOWN;
+};
+
+template<>
+struct db_type<integer> {
+	static constexpr int d = DB_LONG_LONG;
+};
+
+template<>
+struct db_type<real> {
+	static constexpr int d = DB_DOUBLE;
+};
+
+template<>
+struct db_type<char> {
+	static constexpr int d = DB_CHAR;
+};
+
+template<>
+struct db_type<std::int8_t> {
+	static constexpr int d = DB_CHAR;
+};
+
+constexpr int db_type<integer>::d;
+constexpr int db_type<char>::d;
+constexpr int db_type<real>::d;
+
+
+static inline std::string oct_to_str(node_location::node_id n) {
+	return hpx::util::format("{:llo}", n);
+}
+
+
+
+static inline std::string outflow_name(const std::string& varname) {
+	return varname + std::string("_outflow");
+}
+
+
+
+int& silo_epoch();
+double& silo_output_time();
+double& silo_output_rotation_time();
+
 
 #endif /* SRC_SILO_HPP_ */
