+++ src/options.cpp
@@ -72,6 +72,7 @@ bool options::process_options(int argc, char *argv[]) {
 	("silo_offset_z", po::value<integer>(&(opts().silo_offset_z))->default_value(0), "")      //
 	("old_amrbnd", po::value<bool>(&(opts().old_amrbnd))->default_value(false), "use old amr boundary interpolation")           //
 	("amrbnd_order", po::value<integer>(&(opts().amrbnd_order))->default_value(1), "amr boundary interpolation order")        //
+	("silo_num_groups", po::value<integer>(&(opts().silo_num_groups))->default_value(1), "Number of SILO I/O groups")        //
 	("core_refine", po::value<bool>(&(opts().core_refine))->default_value(false), "refine cores by one more level")           //
 	("accretor_refine", po::value<integer>(&(opts().accretor_refine))->default_value(0), "number of extra levels for accretor") //
 	("extra_regrid", po::value<integer>(&(opts().extra_regrid))->default_value(0), "number of extra regrids on startup") //
@@ -168,6 +169,11 @@ bool options::process_options(int argc, char *argv[]) {
 		}
 		std::cout << '\n';
 		const auto num_loc = hpx::find_all_localities().size();
+		if (silo_num_groups > num_loc) {
+			printf("Number of SILO file groups cannot be greater than number of localities. Setting silo_num_groupds to %li\n", num_loc);
+			silo_num_groups = num_loc;
+		}
+		SHOW(silo_num_groups);
 		SHOW(old_amrbnd);
 		SHOW(amrbnd_order);
 		SHOW(angmom);
